init commit

+++ src/main/java/com/KLTN/nguyen/hotelbooking/dto/request/BookingRequest.java
@@ -8,6 +8,6 @@ import java.time.LocalDate;
 public class BookingRequest {
     private LocalDate startDate;
     private LocalDate endDate;
-    private Long roomId;
-    private Long paymentMethodId;
+    private Integer roomId;
+    private Integer paymentMethodId;
 }
